Introduce generic fallback settings for ReportSenders
The idea is to drop the "reportSender-" prefix followed by the type
of reporting (as specified in the reporting data) in favor of just
the reporting type as there's issue with the '-' character as options
in such a section cannot be set via environment variable.

This fallback settings provides generic mechanism where ReportSender
implementation can specify list and order of sections that should be
considered when trying to access specific option.